move pyo3test implementation into own module
